requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
courlan==0.6.0
cssselect==1.1.0
dateparser==1.1.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
htmldate==1.0.0
idna==3.3
interchange==2021.0.4
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jusText==3.0.0
langcodes==3.3.0
lxml==4.6.4
matplotlib-inline==0.1.3
monotonic==1.6
nest-asyncio==1.5.1
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pansi==2020.7.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py2neo==2021.2.3
py4j==0.10.9.2
Pygments==2.10.0
pyparsing==3.0.6
pyspark==3.2.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
readability-lxml==0.8.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tld==0.12.6
tornado==6.1
trafilatura==1.0.0
traitlets==5.1.1
tweepy==4.4.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
